+++ src/pages/api/auth/[...nextauth].js
@@ -2,7 +2,7 @@ import NextAuth from "next-auth";
 import CredentialsProvider from "next-auth/providers/credentials";
 import bcrypt from "bcrypt";
 import { getUser } from "../../../server/mongodb/actions/User";
-import urls from "../../../../utils/urls";
+import urls from "../../../server/utils/urls";
 import connectMongo from "../../../server/mongodb/connectMongo";
 
 export const authOptions = {
