+++ src/pages/api/get_users.js
@@ -1,5 +1,5 @@
 import User from "../../server/mongodb/models/User";
-import requestWrapper from "../../../utils/middleware";
+import requestWrapper from "../../server/utils/middleware";
 
 async function handler(req, res) {
   const users = await User.find({}).populate("tenures").sort({ firstName: 1, lastName: 1 });
