+++ src/pages/api/update_member.js
@@ -2,7 +2,7 @@ import { getToken } from "next-auth/jwt";
 import { upsertTenure } from "../../server/mongodb/actions/Tenure";
 import { updateUser, addTenure, createUser, getUser, changeEmail } from "../../server/mongodb/actions/User";
 import { deleteEmail, deleteNewEmail } from "../../server/mongodb/actions/EmailVerification";
-import requestWrapper from "../../../utils/middleware";
+import requestWrapper from "../../server/utils/middleware";
 import { sendEmailVerification } from "../../server/utils/emailFunctions";
 
 async function handler(req, res) {
