+++ src/pages/exec.jsx
@@ -1,6 +1,6 @@
 import Exec from "../screens/Exec";
 import { getSession } from "next-auth/react";
-import { baseAwsUrl } from "../../utils/awsConfig";
+import { baseAwsUrl } from "../server/utils/awsConfig";
 
 export async function getServerSideProps(context) {
   const session = await getSession(context);
