+++ src/screens/Exec/UserTable.tsx
@@ -2,9 +2,9 @@ import React, { useEffect, useState, useMemo } from "react";
 import CircularProgress from "@mui/material/CircularProgress";
 import PaginationTable from "./PaginationTable";
 import TableContext from "../../../utils/contexts/TableContext";
-import { getCurrSemesterYear, sortTenures } from "../../../utils/utilFunctions";
+import { getCurrSemesterYear, sortTenures } from "../../server/utils/memberFunctions";
 import { DBUser, User } from "./types";
-import sendRequest from "../../../utils/sendToBackend";
+import sendRequest from "../../server/utils/sendToBackend";
 
 function UserTable({ currentSemester, newMembers, clearNewMembers, setSemester, semesters, setSemesters, filter }) {
   const [userList, setUserList] = useState<User[]>([]);
